fix: move functions to own file
